Add vidscraper-cmd.
This adds the beginnings of a vidscraper-cmd. It supports auto-scraping
videos, but nothing else, yet. However, the code is written such that
adding support for other actions is easy to do.

The video subcommand supports specifying fields to be returned and also
api keys (though the latter is untested).

It returns data in JSON format--which required adding ``items()`` and
``to_json()`` methods to the Video class.

Also adds a new topic in the documentation.

Closes #6.